+++ application/tests/models/Tracker/Sites/Batoto_test.php
@@ -19,4 +19,9 @@ class Batoto_test extends SiteTestCase {
 	public function test_failure() {
 		$this->_testSiteFailure('Bad Status Code (404)', '00000:--:bad_lang');
 	}
+
+	public function test_custom() {
+		$this->skipTravis('We need login data to make this work properly which we can\'t provide to Travis');
+		$this->_testSiteCustom();
+	}
 }