+++ application/tests/models/Tracker/Sites/DemonicScans_test.php
@@ -11,4 +11,8 @@ class DemonicScans_test extends SiteTestCase {
 	public function test_failure() {
 		$this->markTestSkipped('DemonicScans no longer exists.');
 	}
+
+	public function test_custom() {
+		$this->_testSiteCustom();
+	}
 }