+++ application/tests/models/Tracker/Sites/HotChocolateScans_test.php
@@ -16,4 +16,8 @@ class HotChocolateScans_test extends SiteTestCase {
 	public function test_failure() {
 		$this->_testSiteFailure('Bad Status Code (404)');
 	}
+
+	public function test_custom() {
+		$this->_testSiteCustom();
+	}
 }