+++ application/tests/models/Tracker/Sites/PureMashiroScans_test.php
@@ -11,11 +11,16 @@ class PureMashiroScans_test extends SiteTestCase {
 			'nanoha-yougashiten-no-ii-shigoto-eng' => 'Nanoha Yougashiten no Ii Shigoto [ENG]',
 			'hanebado-eng'                         => 'Hanebado! [ENG]',
 			'ookumo-chan-flashback'                => 'Ookumo-chan Flashback',
-			'100-man'                              => '100-man no Inochi no Ue ni Ore wa Tatteiru [ENG]',
+			'100-man'                              => '100-man no Inochi no Ue ni Ore wa Tatteiru [ENG]',
+
 		];
 		$this->_testSiteSuccessRandom($testSeries);
 	}
 	public function test_failure() {
 		$this->_testSiteFailure('Bad Status Code (404)');
+	}
+
+	public function test_custom() {
+		$this->_testSiteCustom();
 	}
 }