+++ application/tests/models/Tracker/Sites/S2Scans_test.php
@@ -17,4 +17,8 @@ class S2Scans_test extends SiteTestCase {
 	public function test_failure() {
 		$this->_testSiteFailure('Bad Status Code (404)');
 	}
+
+	public function test_custom() {
+		$this->_testSiteCustom();
+	}
 }