+++ application/tests/models/Tracker/Sites/SenseScans_test.php
@@ -18,4 +18,8 @@ class SenseScans_test extends SiteTestCase {
 	public function test_failure() {
 		$this->_testSiteFailure('Bad Status Code (404)');
 	}
+
+	public function test_custom() {
+		$this->_testSiteCustom();
+	}
 }