+++ application/tests/models/Tracker/Sites/WhiteoutScans_test.php
@@ -15,4 +15,8 @@ class WhiteoutScans_test extends SiteTestCase {
 	public function test_failure() {
 		$this->_testSiteFailure('Bad Status Code (404)');
 	}
+
+	public function test_custom() {
+		$this->_testSiteCustom();
+	}
 }